ml-inference/requirements.txt
fastapi[standard]
requests

biopython==1.85
fair-esm==2.0.0
filelock==3.13.1
fsspec==2023.10.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
torch==2.6.0
typing_extensions>=4.10.0

biotite